+++ src/main/java/com/idan/coupons/controller/CustomerController.java
@@ -2,6 +2,8 @@ package com.idan.coupons.controller;
 
 import java.util.ArrayList;
 import java.util.List;
+
+//import com.idan.coupons.beans.Company;
 import com.idan.coupons.beans.Customer;
 //import com.idan.coupons.dao.CompanyDao;
 //import com.idan.coupons.dao.CouponDao;
@@ -24,6 +26,11 @@ public class CustomerController {
 		this.customerDao=new CustomerDao();
 	}
 	
+	/**
+	 * Creating a customer in the DB.
+	 * @param customer - the customer as a Customer object to add to the DB.
+	 * @throws ApplicationException
+	 */
 	public void createCustomer(Customer customer) throws ApplicationException {
 		
 		validateCreateCustomer(customer);
@@ -32,13 +39,26 @@ public class CustomerController {
 		this.customerDao.createCustomer(customer);
 	}
 	
+	/**
+	 * Removing customer from customer table.
+	 * @param customerID - a long parameter represent the ID of the requested customer.
+	 * @throws ApplicationException
+	 */
 	public void removeCustomerByCustomerID(Long customerID) throws ApplicationException {
-		
+		if(customerID==null) {
+			throw new ApplicationException(ErrorType.BAD_INPUT, DateUtils.getCurrentDateAndTime()
+					+"  Bad input inserted, null value.");
+		}
 //		this.couponDao.removeCustomerPurchasesByCustomerID(customerID);
 		this.customerDao.removeCustomerByCustomerID(customerID);
 		
 	}
 	
+	/**
+	 * Updating a customer in the customer table. All the fields will be updated according to the ID of the customer object.
+	 * @param customer
+	 * @throws ApplicationException
+	 */
 	public void updateCustomer(Customer customer) throws ApplicationException {
 		
 		validateUpdateCustomer(customer);
@@ -48,17 +68,75 @@ public class CustomerController {
 		
 	}
 	
-	public Customer getCustomerByCustomerId(Long customerId) throws ApplicationException {
-		
-		Customer customer = this.customerDao.getCustomerByCustomerId(customerId);
+	/**
+	 * Getting information of a customer.
+	 * @param customerID - a long parameter represent the ID of the requested customer.
+	 * @return Customer object of the requested customer.
+	 * @throws ApplicationException
+	 */
+	public Customer getCustomerByCustomerId(Long customerID) throws ApplicationException {
+		if(customerID==null) {
+			throw new ApplicationException(ErrorType.BAD_INPUT, DateUtils.getCurrentDateAndTime()
+					+"  Bad input inserted, null value.");
+		}
+		Customer customer = this.customerDao.getCustomerByCustomerId(customerID);
 		if (customer == null) {
 			throw new ApplicationException(ErrorType.NO_RETURN_OBJECT, DateUtils.getCurrentDateAndTime()
-					+" No customer with ID" + customerId + ".");
+					+" No customer with ID" + customerID + ".");
 		}
 		
 		return customer;
 	}
 	
+	/**
+	 * Getting information of a customer by name.
+	 * @param customerName a String parameter represent the name of the requested customer.
+	 * @return List of customers object of the requested customer name.
+	 * @throws ApplicationException
+	 */
+	public List<Customer> getCustomersByCustomerName(String customerName) throws ApplicationException {
+		if (!ValidationUtils.isValidNameFormat(customerName)) {
+			throw new ApplicationException(ErrorType.INVALID_PARAMETER, DateUtils.getCurrentDateAndTime()
+					+" Not valid name format" + customerName + ".");
+		}
+		
+		List<Customer> customers = customerDao.getCustomersByCustomerName(customerName);
+		
+		if (customers.isEmpty()) {
+			throw new ApplicationException(ErrorType.NO_RETURN_OBJECT, DateUtils.getCurrentDateAndTime()
+					+" No customer with name" + customerName + ".");
+		}
+		
+		return customers;
+	}
+
+	/**
+	 * Getting information of a customer by e-mail.
+	 * @param customerEmail - a String parameter represent the e-mail of the requested customer.
+	 * @return Customer object of the requested customer.
+	 * @throws ApplicationException
+	 */
+	public Customer getCustomerByCustomerEmail(String customerEmail) throws ApplicationException {
+		if (!ValidationUtils.isValidEmailFormat(customerEmail)) {
+			throw new ApplicationException(ErrorType.INVALID_PARAMETER, DateUtils.getCurrentDateAndTime()
+					+" Not valid name format" + customerEmail + ".");
+		}
+		
+		Customer customer = customerDao.getCustomerByCustomerEmail(customerEmail);
+		
+		if (customer == null) {
+			throw new ApplicationException(ErrorType.NO_RETURN_OBJECT, DateUtils.getCurrentDateAndTime()
+					+" No customer with name" + customerEmail + ".");
+		}
+		
+		return customer;
+	}
+
+	/**
+	 * Getting list of all customers in DB.
+	 * @return List collection of all the customers in the customer table.
+	 * @throws ApplicationException
+	 */
 	public List<Customer> getAllCustomers() throws ApplicationException{
 		
 		List<Customer> customers = this.customerDao.getAllCustomers();
@@ -71,14 +149,26 @@ public class CustomerController {
 		return customers;
 		
 	}
-	
-	public boolean login (String customerEmail, String customerPassword) throws ApplicationException {
+
+	/**
+	 * Logging in to web site.
+	 * @param customerEmail - String of the customer email.
+	 * @param customerPassword - String of that customer password.
+	 * @return The company object that fits the parameters.
+	 * @throws ApplicationException
+	 */
+	public Customer login (String customerEmail, String customerPassword) throws ApplicationException {
 		
 		validateCustomer(new Customer("Valid Name", customerPassword, customerEmail));
 		
 		return this.customerDao.login(customerEmail, customerPassword);
 	}
 
+	/**
+	 * Validating company object for update. In case of invalid parameter, ApplicationException will be thrown.
+	 * @param customer - customer object to validate.
+	 * @throws ApplicationException
+	 */
 	private void validateUpdateCustomer(Customer customer) throws ApplicationException {
 
 		validateCustomer(customer);
@@ -92,6 +182,11 @@ public class CustomerController {
 		
 	}
 
+	/**
+	 * Validating customer object for update. In case of invalid parameter, ApplicationException will be thrown.
+	 * @param customer - customer object to validate.
+	 * @throws ApplicationException
+	 */
 	private void validateCreateCustomer(Customer customer) throws ApplicationException {
 
 		validateCustomer(customer);
@@ -105,9 +200,14 @@ public class CustomerController {
 		
 	}
 
+	/**
+	 * Validating customer object for creation. In case of invalid parameter, ApplicationException will be thrown
+	 * @param customer - customer object to validate.
+	 * @throws ApplicationException
+	 */
 	private void validateCustomer(Customer customer) throws ApplicationException {
 
-		List<InputErrorType> errorTypes = new ArrayList<InputErrorType>();
+		List<InputErrorType> errorTypes = new ArrayList<>();
 		
 		if(!ValidationUtils.isValidNameFormat(customer.getCustomerName())) {
 			errorTypes.add(InputErrorType.INVALID_NAME);
